# backend/app-store/src/client/openapi.yaml
...
openapi: 3.0.3
info:
  title: app-store API
  version: 1.0-SNAPSHOT
servers:
- url: http://localhost:30001
paths:
...
  /product/getById/id/{id}:
    get:
      tags:
      - Product Api
      description: get product by id
      operationId: getProductById
      parameters:
      - name: id
        in: path
        required: true
        schema:
          format: int64
          type: integer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  thumbnail:
                    type: string
                  pictures:
                    type: string
                  productCategory:
                    $ref: '#/components/schemas/ProductCategory'
                  name:
                    type: string
                  producer:
                    type: string
                  guarantee:
                    type: string
                  model:
                    type: string
                  price:
                    format: double
                    type: number
                  quantity:
                    format: int32
                    type: integer
                  productOrders:
                    uniqueItems: true
                    type: array
                    items:
                      $ref: '#/components/schemas/ProductOrder'
                  charger:
                    $ref: '#/components/schemas/Charger'
                  cooler:
                    $ref: '#/components/schemas/Cooler'
                  dramMemory:
                    $ref: '#/components/schemas/DRAMMemory'
                  graphicCard:
                    $ref: '#/components/schemas/GraphicCard'
                  hardDrive:
                    $ref: '#/components/schemas/HardDrive'
                  motherboard:
                    $ref: '#/components/schemas/Motherboard'
                  pcCase:
                    $ref: '#/components/schemas/PcCase'
                  processor:
                    $ref: '#/components/schemas/Processor'
                  rate:
                    format: float
                    type: number
                  thumbnailAsByte:
                    type: string
                  picturesAsBytes:
                    type: string
...
  /store/addOpinion:
    post:
      tags:
      - Application Api
      description: add opinion to product
      operationId: addOpinion
      parameters:
      - name: opinion
        in: query
        schema:
          type: string
      - name: productId
        in: query
        schema:
          format: int64
          type: integer
      - name: rate
        in: query
        schema:
          format: float
          type: number
      responses:
        "204":
          description: NO CONTENT
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /store/addProductToBucket:
    post:
      tags:
      - Application Api
      description: add product to active bucket
      operationId: addProductToBucket
      parameters:
      - name: productId
        in: query
        schema:
          format: int64
          type: integer
      - name: quantity
        in: query
        schema:
          format: int32
          type: integer
      responses:
        "204":
          description: NO CONTENT
        "304":
          description: NOT MODIFIED
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /store/create/order:
    post:
      tags:
      - Application Api
      description: create new order
      operationId: createOrder
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  ordered:
                    $ref: '#/components/schemas/Date'
                  delivered:
                    $ref: '#/components/schemas/Date'
                  status:
                    $ref: '#/components/schemas/Status'
                  totalPrice:
                    format: double
                    type: number
                  bucketId:
                    format: int64
                    type: integer
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /store/create/person:
    post:
      tags:
      - Application Api
      description: create new person
      operationId: createPerson
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Person'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  username:
                    type: string
                  name:
                    type: string
                  lastName:
                    type: string
                  addresses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AddressDTO'
  /store/removeOpinion:
    post:
      tags:
      - Application Api
      description: remove opinion from product
      operationId: removeOpinion
      parameters:
      - name: opinionId
        in: query
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: NO CONTENT
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /store/removeProductFromBucket:
    post:
      tags:
      - Application Api
      description: remove product from bucket
      operationId: removeProductFromBucket
      parameters:
      - name: productId
        in: query
        schema:
          format: int64
          type: integer
      responses:
        "204":
          description: NO CONTENT
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
  /store/updatePersonAddress:
    patch:
      tags:
      - Application Api
      description: "update person address by person id, if address not exists then\
        \ create new one"
      operationId: updatePersonAddressById
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Address'
      responses:
        "202":
          description: ACCEPTED
          content:
            application/json:
              schema:
                type: object
                properties:
                  bid:
                    format: int64
                    type: integer
                  username:
                    type: string
                  name:
                    type: string
                  lastName:
                    type: string
                  addresses:
                    type: array
                    items:
                      $ref: '#/components/schemas/AddressDTO'
        "401":
          description: Not Authorized
        "403":
          description: Not Allowed
      security:
      - SecurityScheme:
        - USER
      - SecurityScheme:
        - USER
components:
  schemas:
    Address:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        province:
          type: string
        zipCode:
          type: string
        streetName:
          type: string
        buildingNumber:
          type: string
        apartmentNumber:
          format: int32
          type: integer
    AddressDTO:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        province:
          type: string
        zipCode:
          type: string
        streetName:
          type: string
        buildingNumber:
          type: string
        apartmentNumber:
          format: int32
          type: integer
...
    Charger:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        power:
          type: string
        standard:
          type: string
        noise:
          format: int32
          type: integer
        coolingType:
          type: string
        width:
          format: int32
          type: integer
        height:
          format: int32
          type: integer
        depth:
          format: int32
          type: integer
...
    Cooler:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        type:
          type: string
        maxRotationSpeed:
          format: int32
          type: integer
        maxVolume:
          format: int32
          type: integer
        supplyVoltage:
          format: int32
          type: integer
        coolerType:
          type: string
...
    DRAMMemory:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        latencyCycle:
          type: string
        memory:
          format: int32
          type: integer
        memoryType:
          type: string
        frequency:
          type: string
...
    Date:
      format: date
      type: string
      example: 2022-03-10
    GraphicCard:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        memoryChipset:
          type: string
        memory:
          format: int32
          type: integer
        connectorType:
          type: string
        memoryType:
          type: string
        memoryClocking:
          type: string
...
    HardDrive:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        memoryInterface:
          type: string
        memory:
          format: int32
          type: integer
        memoryType:
          type: string
...
    Motherboard:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        motherboardStandard:
          type: string
        processorSocket:
          type: string
        memoryType:
          type: string
        maxMemory:
          format: int32
          type: integer
...
    PcCase:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        width:
          format: double
          type: number
        length:
          format: double
          type: number
        depth:
          format: double
          type: number
...
    Person:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        username:
          type: string
        password:
          type: string
        email:
          type: string
        role:
          $ref: '#/components/schemas/Role'
        name:
          type: string
        lastName:
          type: string
        addresses:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/Address'
...
    Processor:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        processorType:
          type: string
        socketType:
          type: string
        numberOfCores:
          format: int32
          type: integer
        numberOfThreads:
          format: int32
          type: integer
        l3Capacity:
          format: int32
          type: integer
...
    Product:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        thumbnail:
          type: string
        pictures:
          type: string
        productCategory:
          $ref: '#/components/schemas/ProductCategory'
        name:
          type: string
        producer:
          type: string
        guarantee:
          type: string
        model:
          type: string
        price:
          format: double
          type: number
        quantity:
          format: int32
          type: integer
        productOrders:
          uniqueItems: true
          type: array
          items:
            $ref: '#/components/schemas/ProductOrder'
        charger:
          $ref: '#/components/schemas/Charger'
        cooler:
          $ref: '#/components/schemas/Cooler'
        dramMemory:
          $ref: '#/components/schemas/DRAMMemory'
        graphicCard:
          $ref: '#/components/schemas/GraphicCard'
        hardDrive:
          $ref: '#/components/schemas/HardDrive'
        motherboard:
          $ref: '#/components/schemas/Motherboard'
        pcCase:
          $ref: '#/components/schemas/PcCase'
        processor:
          $ref: '#/components/schemas/Processor'
        rate:
          format: float
          type: number
        thumbnailAsByte:
          type: string
        picturesAsBytes:
          type: string
    ProductCategory:
      enum:
      - PROCESSOR
      - GRAPHIC_CARD
      - DRAM_MEMORY
      - MOTHERBOARD
      - COOLER
      - PC_CASE
      - HARD_DRIVE
      - CHARGER
      type: string
...
    ProductOrder:
      type: object
      properties:
        bid:
          format: int64
          type: integer
        quantityProductOrder:
          format: int32
          type: integer
        product:
          $ref: '#/components/schemas/Product'
...
    Role:
      enum:
      - ADMIN
      - USER
      - NOT_LOGGED
      type: string
    Status:
      enum:
      - IN_PROGRESS
      - DELIVERED
      - ON_THE_WAY
      type: string